Implemented Dirichlet boundary conditions
Code can now handle inhomgeneous dirichlet boundary conditions, both in
exact and in fem computation. To this end a separate Solution class was
made, which takes as input the stiffness, source, and natural boundary
objects, and applies the dirichlet bc before calculating the solution.

The matrix form of the diffusion and reaction operators is now calculated
in the initialization of the stiffness object, and saved in the
diffusion and reaction operators. Similar for the natural boundary
conditions.